+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2023-10-16
 
 ### Changed
+- used utils.get_colored_logger in other python files
 
 ## [0.3.0] - 2023-10-13